Implemented Sensor model and probability calcukations
Implemented probability calculations and sensor model.  It takes the sensor reading and calculates how many cells occupy that distance then using a bit of trigonometry it figures out what cells lie within the cone the sensor produces (Assumed to have an angle of 40 degrees). Since there are big assumptions on what cells are within the cone multiple scans are advised.